Move tools.(cpp|h) from src/zimwriterfs/ to src/ so the other tools can use it.
