+++ src/tools.h
