+++ src/zimwriterfs/article.cpp
@@ -19,7 +19,7 @@
  */
 
 #include "article.h"
-#include "tools.h"
+#include "../tools.h"
 
 #include <iomanip>
 #include <sstream>
