+++ src/zimwriterfs/meson.build
@@ -40,7 +40,7 @@ endif
 
 sources = [
   'zimwriterfs.cpp',
-  'tools.cpp',
+  '../tools.cpp',
   'article.cpp',
   'zimcreatorfs.cpp',
   'mimetypecounter.cpp'
