+++ src/zimwriterfs/zimwriterfs.cpp
@@ -33,7 +33,7 @@
 #include "zimcreatorfs.h"
 #include "mimetypecounter.h"
 #include "queue.h"
-#include "tools.h"
+#include "../tools.h"
 
 /* Check for version number */
 #ifndef VERSION
